pyproject.toml
[project]
name = "long-term-memory-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.21.1",
    "google-generativeai>=0.8.5",
    "weaviate-client>=4.15.2",
]
